% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgg_collection.R
\name{bgg_collection}
\alias{bgg_collection}
\title{Games in a user's collection by username.}
\usage{
bgg_collection(username, include_collection_expansions = TRUE)
}
\arguments{
\item{username}{String. BoardGameGeek user's username.}

\item{include_collection_expansions}{Boolean. Whether or not to include expansions as part of the user's collection. Set to TRUE by default.}
}
\description{
\code{bgg_collection} Returns all games in a user's collection, as a tibble.
}
